Move wishlist related components under 'product' package
