+++ src/main/java/com/github/jbence1994/webshop/product/WishlistProductDto.java
@@ -1,0 +1,4 @@
+package com.github.jbence1994.webshop.product;
+
+public record WishlistProductDto(Long id) {
+}
