+++ src/main/java/com/github/jbence1994/webshop/product/WishlistService.java
@@ -1,6 +1,4 @@
-package com.github.jbence1994.webshop.wishlist;
-
-import com.github.jbence1994.webshop.product.Product;
+package com.github.jbence1994.webshop.product;
 
 public interface WishlistService {
     Product addProductToWishlist(Long productId);
