--- src/main/java/com/github/jbence1994/webshop/wishlist/WishlistProductDto.java
@@ -1,4 +1,0 @@
-package com.github.jbence1994.webshop.wishlist;
-
-public record WishlistProductDto(Long id) {
-}
